requirements.txt
# Project dependencies
aiohttp>=3.9.0
python-dateutil>=2.8.2
pydantic>=2.4.0

openai>=1.0.0
python-dotenv>=1.0.0

# AuthService dependencies
fastapi==0.115.0
uvicorn[standard]==0.30.0
asyncpg==0.29.0
pydantic==2.7.0
pydantic[email]
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
aioredis==2.0.1
twilio==9.0.0
aiosmtplib==3.0.1
python-dotenv==1.0.0

# Test dependencies
pytest>=7.4.0
pytest-asyncio>=0.23.0
httpx>=0.24.0